requirements.txt
openai
requests
pyttsx3
SpeechRecognition
psutil
python-dotenv